quantile and threshold color scales
`{type: "threshold", domain: [30, 50], range: ["red", "green", "blue"]}`
closes #372

`color: {type: "quantile", scheme: "blues", quantiles: 7}`
closes #373
