+++ src/scales/quantitative.js
@@ -1,53 +1,15 @@
 import {
-  min,
-  max,
-  quantile,
-  reverse as reverseof,
-  piecewise,
   interpolateHcl,
   interpolateHsl,
   interpolateLab,
   interpolateNumber,
   interpolateRgb,
   interpolateRound,
-  interpolateBlues,
-  interpolateBrBG,
-  interpolateBuGn,
-  interpolateBuPu,
-  interpolateCividis,
-  interpolateCool,
-  interpolateCubehelixDefault,
-  interpolateGnBu,
-  interpolateGreens,
-  interpolateGreys,
-  interpolateInferno,
-  interpolateMagma,
-  interpolateOranges,
-  interpolateOrRd,
-  interpolatePiYG,
-  interpolatePlasma,
-  interpolatePRGn,
-  interpolatePuBu,
-  interpolatePuBuGn,
-  interpolatePuOr,
-  interpolatePuRd,
-  interpolatePurples,
-  interpolateRainbow,
-  interpolateRdBu,
-  interpolateRdGy,
-  interpolateRdPu,
-  interpolateRdYlBu,
-  interpolateRdYlGn,
-  interpolateReds,
-  interpolateSinebow,
-  interpolateSpectral,
-  interpolateTurbo,
-  interpolateViridis,
-  interpolateWarm,
-  interpolateYlGn,
-  interpolateYlGnBu,
-  interpolateYlOrBr,
-  interpolateYlOrRd,
+  min,
+  max,
+  quantile,
+  reverse as reverseof,
+  piecewise,
   scaleDiverging,
   scaleDivergingLog,
   scaleDivergingPow,
@@ -57,16 +19,18 @@ import {
   scaleLog,
   scalePow,
   scaleSqrt,
+  scaleQuantile,
   scaleSymlog,
+  scaleThreshold,
   scaleIdentity
 } from "d3";
+import {ordinalScheme, quantitativeScheme, quantitativeSchemes} from "./schemes.js";
 import {registry, radius, opacity, color} from "./index.js";
-import {positive, negative} from "../defined.js";
+import {defined, positive, negative} from "../defined.js";
 import {constant} from "../mark.js";
 
 const flip = i => t => i(1 - t);
 
-// TODO Allow this to be extended.
 const interpolators = new Map([
   // numbers
   ["number", interpolateNumber],
@@ -78,69 +42,10 @@ const interpolators = new Map([
   ["lab", interpolateLab]
 ]);
 
-// TODO Allow this to be extended.
-const schemes = new Map([
-  // diverging
-  ["brbg", interpolateBrBG],
-  ["prgn", interpolatePRGn],
-  ["piyg", interpolatePiYG],
-  ["puor", interpolatePuOr],
-  ["rdbu", interpolateRdBu],
-  ["rdgy", interpolateRdGy],
-  ["rdylbu", interpolateRdYlBu],
-  ["rdylgn", interpolateRdYlGn],
-  ["spectral", interpolateSpectral],
-
-  // reversed diverging (for temperature data)
-  ["burd", t => interpolateRdBu(1 - t)],
-  ["buylrd", t => interpolateRdYlBu(1 - t)],
-
-  // sequential (single-hue)
-  ["blues", interpolateBlues],
-  ["greens", interpolateGreens],
-  ["greys", interpolateGreys],
-  ["purples", interpolatePurples],
-  ["reds", interpolateReds],
-  ["oranges", interpolateOranges],
-
-  // sequential (multi-hue)
-  ["turbo", interpolateTurbo],
-  ["viridis", interpolateViridis],
-  ["magma", interpolateMagma],
-  ["inferno", interpolateInferno],
-  ["plasma", interpolatePlasma],
-  ["cividis", interpolateCividis],
-  ["cubehelix", interpolateCubehelixDefault],
-  ["warm", interpolateWarm],
-  ["cool", interpolateCool],
-  ["bugn", interpolateBuGn],
-  ["bupu", interpolateBuPu],
-  ["gnbu", interpolateGnBu],
-  ["orrd", interpolateOrRd],
-  ["pubugn", interpolatePuBuGn],
-  ["pubu", interpolatePuBu],
-  ["purd", interpolatePuRd],
-  ["rdpu", interpolateRdPu],
-  ["ylgnbu", interpolateYlGnBu],
-  ["ylgn", interpolateYlGn],
-  ["ylorbr", interpolateYlOrBr],
-  ["ylorrd", interpolateYlOrRd],
-
-  // cyclical
-  ["rainbow", interpolateRainbow],
-  ["sinebow", interpolateSinebow]
-]);
-
 function Interpolator(interpolate) {
   const i = (interpolate + "").toLowerCase();
   if (!interpolators.has(i)) throw new Error(`unknown interpolator: ${i}`);
   return interpolators.get(i);
-}
-
-function Scheme(scheme) {
-  const s = (scheme + "").toLowerCase();
-  if (!schemes.has(s)) throw new Error(`unknown scheme: ${s}`);
-  return schemes.get(s);
 }
 
 export function ScaleQ(key, scale, channels, {
@@ -153,7 +58,7 @@ export function ScaleQ(key, scale, channels, {
   range = registry.get(key) === radius ? inferRadialRange(channels, domain) : registry.get(key) === opacity ? [0, 1] : undefined,
   scheme,
   type,
-  interpolate = registry.get(key) === color ? (range !== undefined ? interpolateRgb : scheme !== undefined ? Scheme(scheme) : type === "cyclical" ? interpolateRainbow : interpolateTurbo) : round ? interpolateRound : undefined,
+  interpolate = registry.get(key) === color ? (range !== undefined ? interpolateRgb : scheme !== undefined ? quantitativeScheme(scheme) : quantitativeSchemes.get(type === "cyclical" ? "rainbow" : "turbo")) : round ? interpolateRound : undefined,
   reverse,
   inset
 }) {
@@ -167,7 +72,7 @@ export function ScaleQ(key, scale, channels, {
   // is used in conjunction with the range. And other times the interpolate
   // function is a “fixed” interpolator independent of the range, as when a
   // color scheme such as interpolateRdBu is used.
-  if (interpolate !== undefined) {
+  if (scale.interpolate && interpolate !== undefined) {
     if (typeof interpolate !== "function") {
       interpolate = Interpolator(interpolate);
     } else if (interpolate.length === 1) {
@@ -198,8 +103,27 @@ export function ScaleLog(key, channels, {base = 10, domain = inferLogDomain(chan
   return ScaleQ(key, scaleLog().base(base), channels, {domain, ...options});
 }
 
+export function ScaleQuantile(key, channels, {
+  quantiles = 5,
+  scheme,
+  domain = inferFullDomain(channels),
+  range = ordinalScheme(scheme === undefined ? "rdylbu" : scheme)({length: quantiles}).slice(0, quantiles),
+  ...options
+}) {
+  return ScaleQ(key, scaleQuantile(), [], {domain, range, ...options});
+}
+
 export function ScaleSymlog(key, channels, {constant = 1, ...options}) {
   return ScaleQ(key, scaleSymlog().constant(constant), channels, options);
+}
+
+export function ScaleThreshold(key, channels, {
+  domain = inferDomain(channels),
+  scheme,
+  range = ordinalScheme(scheme === undefined ? "rdylbu" : scheme)({length: domain.length + 1}),
+  ...options
+}) {
+  return ScaleQ(key, scaleThreshold(), channels, {domain, range, ...options});
 }
 
 export function ScaleIdentity() {
@@ -213,7 +137,7 @@ function ScaleD(key, scale, channels, {
   pivot = 0,
   range,
   scheme,
-  interpolate = registry.get(key) === color ? (range !== undefined ? interpolateRgb : scheme !== undefined ? Scheme(scheme) : interpolateRdBu) : undefined,
+  interpolate = registry.get(key) === color ? (range !== undefined ? interpolateRgb : quantitativeScheme(scheme !== undefined ? scheme : "rdbu")) : undefined,
   reverse
 }) {
   domain = [Math.min(domain[0], pivot), pivot, Math.max(domain[1], pivot)];
@@ -283,4 +207,12 @@ function inferLogDomain(channels) {
     }
   }
   return [1, 10];
+}
+
+function inferFullDomain(channels) {
+  let domain = [];
+  for (const {value} of channels) {
+    if (value !== undefined) domain = domain.concat(value);
+  }
+  return domain.filter(defined);
 }
